rename services/api.js to services/osm-api.js to avoid confusion with Observe API
