+++ app/actions/map.js
@@ -10,7 +10,7 @@ import sizeof from 'object-sizeof'
 import * as types from './actionTypes'
 import { setNotification } from './notification'
 import { getOfflineResources, getOfflineResourceStatus, getPendingEviction } from '../selectors'
-import { saveDataForTile } from '../services/api'
+import { saveDataForTile } from '../services/osm-api'
 import { bboxToTiles } from '../utils/bbox'
 import cache from '../utils/data-cache'
 import { filterRelations } from '../utils/filter-xml'
